requirements.txt
pyserial==3.5
pyyaml==6.0.1
